+++ PortfolioManagerWASM/Components/Home/OverviewPortfoliosComponent.razor.cs
@@ -1,8 +1,5 @@
 using Microsoft.AspNetCore.Components;
-using PortfolioManagerWASM.Helpers;
 using PortfolioManagerWASM.Models;
-using PortfolioManagerWASM.Pages;
-using PortfolioManagerWASM.ViewModels;
 
 namespace PortfolioManagerWASM.Components.Home
 {
@@ -55,6 +52,6 @@ namespace PortfolioManagerWASM.Components.Home
     public class ChartDataModel
     {
         public string AssetName { get; set; }
-        public float Allocation { get; set; }
+        public int Allocation { get; set; }
     }
 }
